feat(E2EI): reduce complexity of OIDCservice, mock data
